+++ x/crosschain/keeper/initiate_outbound.go
@@ -38,6 +38,6 @@ func (k Keeper) InitiateOutbound(ctx sdk.Context, config InitiateOutboundConfig)
 		)
 	}
 
-	config.CCTX.SetPendingOutbound("")
+	config.CCTX.SetPendingOutbound(types.StatusMessages{StatusMessage: "initiating outbound"})
 	return cctxGateway.InitiateOutbound(ctx, config)
 }