+++ zetaclient/chains/bitcoin/bitcoin.go
@@ -11,33 +11,23 @@ import (
 	"github.com/zeta-chain/node/pkg/chains"
 	"github.com/zeta-chain/node/pkg/scheduler"
 	"github.com/zeta-chain/node/pkg/ticker"
+	"github.com/zeta-chain/node/zetaclient/chains/base"
 	"github.com/zeta-chain/node/zetaclient/chains/bitcoin/observer"
 	"github.com/zeta-chain/node/zetaclient/chains/bitcoin/signer"
 	zctx "github.com/zeta-chain/node/zetaclient/context"
-	"github.com/zeta-chain/node/zetaclient/outboundprocessor"
 )
 
 type Bitcoin struct {
 	scheduler *scheduler.Scheduler
 	observer  *observer.Observer
 	signer    *signer.Signer
-	proc      *outboundprocessor.Processor
 }
 
-func New(
-	scheduler *scheduler.Scheduler,
-	observer *observer.Observer,
-	signer *signer.Signer,
-) *Bitcoin {
-	// TODO move this to base signer
-	// https://github.com/zeta-chain/node/issues/3330
-	proc := outboundprocessor.NewProcessor(observer.Logger().Outbound)
-
+func New(scheduler *scheduler.Scheduler, observer *observer.Observer, signer *signer.Signer) *Bitcoin {
 	return &Bitcoin{
 		scheduler: scheduler,
 		observer:  observer,
 		signer:    signer,
-		proc:      proc,
 	}
 }
 
@@ -154,7 +144,7 @@ func (b *Bitcoin) scheduleCCTX(ctx context.Context) error {
 		var (
 			params     = cctx.GetCurrentOutboundParam()
 			nonce      = params.TssNonce
-			outboundID = outboundprocessor.ToOutboundID(cctx.Index, params.ReceiverChainId, nonce)
+			outboundID = base.OutboundIDFromCCTX(cctx)
 		)
 
 		if params.ReceiverChainId != chainID {
@@ -183,18 +173,14 @@ func (b *Bitcoin) scheduleCCTX(ctx context.Context) error {
 				Uint64("outbound.earliest_pending_nonce", cctxList[0].GetCurrentOutboundParam().TssNonce).
 				Msg("Schedule CCTX: lookahead reached")
 			return nil
-		case b.proc.IsOutboundActive(outboundID):
+		case b.signer.IsOutboundActive(outboundID):
 			// outbound is already being processed
 			continue
 		}
 
-		b.proc.StartTryProcess(outboundID)
-
 		go b.signer.TryProcessOutbound(
 			ctx,
 			cctx,
-			b.proc,
-			outboundID,
 			b.observer,
 			b.observer.ZetacoreClient(),
 			zetaHeight,