+++ zetaclient/chains/ton/observer/inbound.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/zeta-chain/node/pkg/coin"
 	toncontracts "github.com/zeta-chain/node/pkg/contracts/ton"
+	"github.com/zeta-chain/node/x/crosschain/types"
 	"github.com/zeta-chain/node/zetaclient/chains/ton/liteapi"
 	"github.com/zeta-chain/node/zetaclient/zetacore"
 )
@@ -263,6 +264,7 @@ func (ob *Observer) voteDeposit(
 		asset,
 		operatorAddress.String(),
 		eventIndex,
+		types.InboundStatus_SUCCESS,
 	)
 
 	return ob.PostVoteInbound(ctx, msg, retryGasLimit)