+++ zetaclient/chains/ton/signer/signer.go
@@ -17,7 +17,6 @@ import (
 	cc "github.com/zeta-chain/node/x/crosschain/types"
 	"github.com/zeta-chain/node/zetaclient/chains/base"
 	"github.com/zeta-chain/node/zetaclient/chains/interfaces"
-	"github.com/zeta-chain/node/zetaclient/outboundprocessor"
 )
 
 // LiteClient represents a TON client
@@ -72,13 +71,12 @@ func New(baseSigner *base.Signer, client LiteClient, gateway *toncontracts.Gatew
 func (s *Signer) TryProcessOutbound(
 	ctx context.Context,
 	cctx *cc.CrossChainTx,
-	proc *outboundprocessor.Processor,
-	outboundID string,
 	zetacore interfaces.ZetacoreClient,
 	zetaBlockHeight uint64,
 ) {
-	proc.StartTryProcess(outboundID)
-	defer proc.EndTryProcess(outboundID)
+	outboundID := base.OutboundIDFromCCTX(cctx)
+	s.MarkOutbound(outboundID, true)
+	defer s.MarkOutbound(outboundID, false)
 
 	outcome, err := s.ProcessOutbound(ctx, cctx, zetacore, zetaBlockHeight)