+++ zetaclient/orchestrator/orchestrator.go
@@ -23,10 +23,10 @@ import (
 	"github.com/zeta-chain/node/zetaclient/chains/base"
 	"github.com/zeta-chain/node/zetaclient/chains/interfaces"
 	solanaobserver "github.com/zeta-chain/node/zetaclient/chains/solana/observer"
+	solanasigner "github.com/zeta-chain/node/zetaclient/chains/solana/signer"
 	zctx "github.com/zeta-chain/node/zetaclient/context"
 	"github.com/zeta-chain/node/zetaclient/logs"
 	"github.com/zeta-chain/node/zetaclient/metrics"
-	"github.com/zeta-chain/node/zetaclient/outboundprocessor"
 	"github.com/zeta-chain/node/zetaclient/ratelimiter"
 )
 
@@ -53,9 +53,6 @@ type Orchestrator struct {
 
 	// observerMap contains the chain observers indexed by chainID
 	observerMap map[int64]interfaces.ChainObserver
-
-	// outbound processor
-	outboundProc *outboundprocessor.Processor
 
 	// last operator balance
 	lastOperatorBalance sdkmath.Int
@@ -110,7 +107,6 @@ func New(
 		signerMap:   signerMap,
 		observerMap: observerMap,
 
-		outboundProc:        outboundprocessor.NewProcessor(logger.Std),
 		lastOperatorBalance: balance,
 
 		// observer & signer props
@@ -429,8 +425,18 @@ func (oc *Orchestrator) ScheduleCctxSolana(
 	signer interfaces.ChainSigner,
 ) {
 	solObserver, ok := observer.(*solanaobserver.Observer)
-	if !ok { // should never happen
+
+	// should never happen
+	if !ok {
 		oc.logger.Error().Msgf("ScheduleCctxSolana: chain observer is not a solana observer")
+		return
+	}
+
+	solSigner, ok := signer.(*solanasigner.Signer)
+
+	// should never happen
+	if !ok {
+		oc.logger.Error().Msgf("ScheduleCctxSolana: chain signer is not a solana signer")
 		return
 	}
 
@@ -444,7 +450,7 @@ func (oc *Orchestrator) ScheduleCctxSolana(
 	for _, cctx := range cctxList {
 		params := cctx.GetCurrentOutboundParam()
 		nonce := params.TssNonce
-		outboundID := outboundprocessor.ToOutboundID(cctx.Index, params.ReceiverChainId, nonce)
+		outboundID := base.OutboundIDFromCCTX(cctx)
 
 		if params.ReceiverChainId != chainID {
 			oc.logger.Error().
@@ -471,15 +477,16 @@ func (oc *Orchestrator) ScheduleCctxSolana(
 			continue
 		}
 
+		// noop
+		if solSigner.IsOutboundActive(outboundID) {
+			continue
+		}
+
 		// schedule a TSS keysign
-		if nonce%interval == zetaHeight%interval && !oc.outboundProc.IsOutboundActive(outboundID) {
-			oc.outboundProc.StartTryProcess(outboundID)
-			oc.logger.Debug().Msgf("ScheduleCctxSolana: sign outbound %s with value %d", outboundID, params.Amount)
+		if nonce%interval == zetaHeight%interval {
 			go signer.TryProcessOutbound(
 				ctx,
 				cctx,
-				oc.outboundProc,
-				outboundID,
 				observer,
 				oc.zetacoreClient,
 				zetaHeight,